# storage: combine spans when adding to the timestamp cache