```
name                  old time/op    new time/op    delta
KVInsert1_SQL-32         504µs ± 2%     508µs ± 2%  +0.94%  (p=0.008 n=19+20)
KVInsert10_SQL-32        956µs ± 2%     934µs ± 2%  -2.34%  (p=0.000 n=20+20)
KVInsert100_SQL-32      5.05ms ± 2%    4.81ms ± 2%  -4.70%  (p=0.000 n=20+18)
KVInsert1000_SQL-32     48.5ms ± 5%    45.0ms ± 3%  -7.09%  (p=0.000 n=19+19)
KVInsert10000_SQL-32     651ms ±10%     647ms ±13%    ~     (p=0.828 n=18+20)

name                  old allocs/op  new allocs/op  delta
KVInsert1_SQL-32           304 ± 0%       302 ± 0%  -0.66%  (p=0.000 n=19+20)
KVInsert10_SQL-32          754 ± 0%       717 ± 0%  -4.96%  (p=0.000 n=19+16)
KVInsert100_SQL-32       5.04k ± 0%     4.65k ± 0%  -7.83%  (p=0.000 n=20+20)
KVInsert1000_SQL-32      48.5k ± 2%     44.4k ± 1%  -8.46%  (p=0.000 n=20+20)
KVInsert10000_SQL-32      565k ± 9%      525k ± 9%  -7.06%  (p=0.002 n=20+20)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/6413)
<!-- Reviewable:end -->