@petermattis 2016-04-29
@tschottdorf This is intended for discussion. See also #6412. Similar to that change, perhaps we should use some heuristics to combine spans.