@petermattis 2016-05-03
Here are updated numbers where old is #6412:

```
name                            old time/op    new time/op    delta
InsertDistinct1_Cockroach-32      2.55ms ± 5%    1.99ms ± 4%  -21.85%  (p=0.000 n=20+20)
InsertDistinct10_Cockroach-32     2.19ms ± 5%    1.50ms ± 4%  -31.31%  (p=0.000 n=20+20)
InsertDistinct100_Cockroach-32    2.56ms ± 6%    2.04ms ± 7%  -20.08%  (p=0.000 n=20+18)
KVInsert1_SQL-32                   503µs ± 2%     495µs ± 2%   -1.51%  (p=0.001 n=19+19)
KVInsert10_SQL-32                  934µs ± 2%     898µs ± 2%   -3.79%  (p=0.000 n=17+20)
KVInsert100_SQL-32                4.79ms ± 4%    4.51ms ± 6%   -5.86%  (p=0.000 n=20+20)
KVInsert1000_SQL-32               42.6ms ± 7%    38.5ms ± 7%   -9.77%  (p=0.000 n=19+18)
KVInsert10000_SQL-32               494ms ±10%     436ms ±10%  -11.77%  (p=0.000 n=18+19)

name                            old allocs/op  new allocs/op  delta
InsertDistinct1_Cockroach-32       2.07k ± 4%     1.95k ± 3%   -6.04%  (p=0.000 n=20+20)
InsertDistinct10_Cockroach-32      2.23k ± 3%     2.11k ± 3%   -5.22%  (p=0.000 n=19+20)
InsertDistinct100_Cockroach-32     4.95k ± 2%     4.88k ± 2%   -1.41%  (p=0.000 n=20+17)
KVInsert1_SQL-32                     294 ± 0%       292 ± 0%   -0.68%  (p=0.000 n=20+19)
KVInsert10_SQL-32                    705 ± 0%       668 ± 0%   -5.25%  (p=0.000 n=20+20)
KVInsert100_SQL-32                 4.62k ± 0%     4.23k ± 0%   -8.54%  (p=0.000 n=20+20)
KVInsert1000_SQL-32                43.8k ± 0%     39.8k ± 0%   -9.14%  (p=0.000 n=17+20)
KVInsert10000_SQL-32                459k ± 0%      419k ± 0%   -8.75%  (p=0.000 n=16+19)
```

So definitely a performance win, but are there correctness or other issues from this change?