@bdarnell 2016-05-03
I think this would be correct, but I would expect performance problems from the wider spans in certain situations, just like in #6412. Since the timestamp cache is mainly there for read/write conflicts, I think a benchmark of transactions that do a bunch of reads (same pattern as InsertDistinct) followed by at least one write would be the best way to expose that performance difference.