@tschottdorf 2016-05-04
Like @bdarnell pointed out, we don't at all exercise the downside of this change in the above benchmarks. The cost we pay for this optimization in practice are transaction restarts under contention. With the exponential backoff we have now, these create a lot of variance (the bank benchmark is the only one we have that actually attempts this, I think) and so we haven't really looked at them much.

A workload which for some reason writes to the beginning and end of a key range in one batch will have all concurrent writers in between restart, which will make some workloads which should be completely conflict-free contended. Even worse, it could lead to starvation. Picture a reader busy-looping a read of keys `a` and `z`, while everyone else is writing random, non-intersecting keys in between. Currently, that's completely conflict-free. With the change, the reader bombs the whole range repeatedly with its read timestamp. All in-flight writers will lay down an intent but have to restart; but when they come back, the timestamp cache again instructs them to restart. They lay down (update) their intent, but when they come back... you get the idea. With this change, we're losing the behavior that laying down an intent forces a concurrent reader into conflict resolution - you're introducing artificial "readers" here which actually never see the intents.

To benchmark changes like this, I think what we might want to consider is "retryless contention" - i.e. we run transactions that give up when they catch a restart or intent error (since a lot of the variance came from internal retry loops). That should give us some measure of how "contended" a workload is (without measuring how much effort a transaction would have to do to succeed).

But, in this case, we have to figure out the starvation scenario outlined above first.

---

Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[roachpb/data.go, line 931 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6413#-KGwlJ9-3fttIyn2eTUd:-KGwlJ9-3fttIyn2eTUe:-2044422802) ([raw file](https://github.com/cockroachdb/cockroach/blob/6b7f1802437795834d06394b6306766a65f02b5d/roachpb/data.go#L931)):*
We should try to avoid relying on `nil != []byte{}` if we can (or `[nil,a)` union `[b, c)` ends up `[b,c)`). I believe all should be well if you just drop this special case (though ultimately the signature of this method suggests that it shouldn't be called on an uninitialized Span at all: `[nil,nil)` could technically stand for a point request to `KeyMin`, but that's not what you mean).
Also, should unit test.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6413#-:-KGwnTX4o6m7Ib6Lpaax:355380824)*
<!-- Sent from Reviewable.io -->