@petermattis 2016-05-12
In most workloads, there isn't conflict. Similar to the `CommandQueue` optimization, I think we can make the `TimestampCache` add covering spans that are split on conflict. For the timestamp cache, conflict is not just when a request encounters a covering span, but also when the new request has a lower timestamp than the covering span. For the `TrackChoices` benchmark, we would never actually split the covering spans as the transaction timestamps are always increasing. I also benchmarked `TrackChoices` against this PR (old) vs a hack to never add to the timestamp cache (new):

```
name                          old time/op    new time/op    delta
TrackChoices1_Cockroach-8        391µs ± 1%     354µs ± 1%  -9.44%          (p=0.008 n=5+5)
TrackChoices10_Cockroach-8       111µs ± 1%     107µs ± 0%  -3.89%          (p=0.008 n=5+5)
TrackChoices100_Cockroach-8     74.8µs ± 1%    72.9µs ± 2%  -2.53%          (p=0.008 n=5+5)
TrackChoices1000_Cockroach-8    77.6µs ± 3%    76.0µs ± 1%    ~             (p=0.222 n=5+5)
```

As expected, completely skipping the timestamp cache is faster for small batch sizes. But at large batch size other effects take over. 

And here is how this PR benchmarks against master:

```
name                          old time/op    new time/op    delta
TrackChoices1_Cockroach-8        445µs ± 2%     391µs ± 1%  -12.24%  (p=0.008 n=5+5)
TrackChoices10_Cockroach-8       143µs ± 1%     111µs ± 1%  -22.32%  (p=0.008 n=5+5)
TrackChoices100_Cockroach-8      112µs ± 0%      75µs ± 1%  -33.07%  (p=0.008 n=5+5)
TrackChoices1000_Cockroach-8     108µs ± 2%      78µs ± 3%  -28.46%  (p=0.008 n=5+5)
```