@petermattis 2016-05-13
This change has been completely revamped. I need to regenerate the performance numbers, but I expect them to be equivalent or better. And there is likely some additional cleanup and testing work to do. Would be good to get some eyes on the general approach, though. PTAL.